Share code.
* New module common.rs where all the shared code lives.
* New trait CollapsePrivate that encapsulates all the code to make
  collapsers work on multiple threads. All types that implement
  CollapsePrivate automatically implement Collapse, but, as the name
  suggests, it is opaque to the end-user (since it's in a private
  module). Implementors of additional collapsers can use it if they'd
  like or they can implement Collapse itself.
* For perf, channel to send event type from main thread to workers was not
  necessary; so I removed it. Instead we just process the first stack
  (where the information on what the event filter should be is) before
  kicking off all the threads.
* `sample` implementation on master does not sort output like all the
  other collapsers do; so I fixed that. This means I had to sort the
  results files we've been testing against as well (that's why they
  changed).
* Also made `sample` binary consistent with all the other binaries.
