+++ Cargo.toml
@@ -49,7 +49,6 @@ libflate = "0.1"
 maplit = "1.0.1"
 pretty_assertions = "0.6"
 rand = "0.6"
-rustversion = "0.1"
 
 [profile.release]
 debug = true
