+++ src/bin/collapse-sample.rs
@@ -1,10 +1,15 @@
-use env_logger::Env;
 use std::io;
 use std::path::PathBuf;
+
+use env_logger::Env;
+use inferno::collapse::sample::{Folder, Options};
+use inferno::collapse::{Collapse, DEFAULT_NTHREADS};
+use lazy_static::lazy_static;
 use structopt::StructOpt;
 
-use inferno::collapse::sample::{Folder, Options};
-use inferno::collapse::Collapse;
+lazy_static! {
+    static ref NTHREADS: String = format!("{}", *DEFAULT_NTHREADS);
+}
 
 #[derive(Debug, StructOpt)]
 #[structopt(
@@ -15,6 +20,9 @@ use inferno::collapse::Collapse;
             sample 1234 -file out.sample_stacks"
 )]
 struct Opt {
+    // ************* //
+    // *** FLAGS *** //
+    // ************* //
     /// Don't include modules with function names
     #[structopt(long = "no-modules")]
     no_modules: bool,
@@ -27,7 +35,23 @@ struct Opt {
     #[structopt(short = "v", long = "verbose", parse(from_occurrences))]
     verbose: usize,
 
+    // *************** //
+    // *** OPTIONS *** //
+    // *************** //
+    /// Number of threads to use.
+    #[structopt(
+        short = "n",
+        long = "nthreads",
+        raw(default_value = "&NTHREADS"),
+        value_name = "UINT"
+    )]
+    nthreads: usize,
+
+    // ************ //
+    // *** ARGS *** //
+    // ************ //
     /// sample output file, or STDIN if not specified
+    #[structopt(value_name = "PATH")]
     infile: Option<PathBuf>,
 }
 
@@ -37,6 +61,7 @@ impl Opt {
             self.infile,
             Options {
                 no_modules: self.no_modules,
+                nthreads: self.nthreads,
             },
         )
     }
