+++ src/collapse/dtrace.rs
@@ -1,18 +1,15 @@
 use std::borrow::Cow;
 use std::collections::VecDeque;
 use std::io::{self, prelude::*};
-use std::mem;
 
-use crossbeam::channel;
 use log::warn;
 
-use crate::collapse::util::fix_partially_demangled_rust_symbol;
-use crate::collapse::{self, Collapse, Occurrences};
+use crate::collapse::common::{self, CollapsePrivate, Occurrences};
 
-/// Dtrace folder configuration options.
+/// `dtrace` folder configuration options.
 #[derive(Clone, Debug)]
 pub struct Options {
-    /// Demangle function names
+    /// Demangle function names. Default is `false`.
     pub demangle: bool,
 
     /// Include function offset (except leafs). Default is `false`.
@@ -27,7 +24,7 @@ impl Default for Options {
         Self {
             demangle: false,
             includeoffset: false,
-            nthreads: *collapse::DEFAULT_NTHREADS,
+            nthreads: *common::DEFAULT_NTHREADS,
         }
     }
 }
@@ -36,15 +33,13 @@ impl Default for Options {
 ///
 /// To construct one, either use `dtrace::Folder::default()` or create an [`Options`] and use
 /// `dtrace::Folder::from(options)`.
+#[derive(Clone)]
 pub struct Folder {
     /// Vector for processing java stuff
     cache_inlines: Vec<String>,
 
-    /// Number of stacks in each job sent to the threadpool.
+    /// The number of stacks per job to send to the threadpool.
     nstacks_per_job: usize,
-
-    /// Number of times each call stack has been seen.
-    occurrences: Occurrences,
 
     /// Function entries on the stack in this entry thus far.
     stack: VecDeque<String>,
@@ -62,8 +57,7 @@ impl From<Options> for Folder {
         }
         Self {
             cache_inlines: Vec::new(),
-            nstacks_per_job: collapse::NSTACKS_PER_JOB,
-            occurrences: Occurrences::new(opt.nthreads),
+            nstacks_per_job: common::DEFAULT_NSTACKS_PER_JOB,
             stack: VecDeque::default(),
             stack_str_size: 0,
             opt,
@@ -77,13 +71,12 @@ impl Default for Folder {
     }
 }
 
-impl Collapse for Folder {
-    fn collapse<R, W>(&mut self, mut reader: R, writer: W) -> io::Result<()>
+impl CollapsePrivate for Folder {
+    fn pre_process<R>(&mut self, reader: &mut R, _: &mut Occurrences) -> io::Result<()>
     where
         R: io::BufRead,
-        W: io::Write,
     {
-        // skip header lines -- first empty line marks start of data
+        // Consumer the header...
         let mut line = String::new();
         loop {
             line.clear();
@@ -93,24 +86,49 @@ impl Collapse for Folder {
                 return Ok(());
             };
             if line.trim().is_empty() {
-                break;
+                return Ok(());
             }
         }
+    }
 
-        // Do collapsing...
-        if self.occurrences.is_concurrent() {
-            self.collapse_multi_threaded(reader)?;
-        } else {
-            self.collapse_single_threaded(reader)?;
+    fn collapse_single_threaded<R>(
+        &mut self,
+        mut reader: R,
+        occurrences: &mut Occurrences,
+    ) -> io::Result<()>
+    where
+        R: io::BufRead,
+    {
+        let mut line = String::new();
+        loop {
+            line.clear();
+            if reader.read_line(&mut line)? == 0 {
+                break;
+            }
+            let line = line.trim();
+            if line.is_empty() {
+                continue;
+            } else if let Ok(count) = line.parse::<usize>() {
+                self.on_stack_end(count, occurrences);
+            } else {
+                self.on_stack_line(line);
+            }
         }
-
-        // Write results...
-        self.occurrences.write_and_clear(writer)?;
-
+        // If we reach this point in the code and there's still something in our
+        // state (`self.stack` and `self.stack_str_size`), it means the input
+        // did not terminate at the end of a stack; rather, it terminated in
+        // the middle of a stack. In this case, we consider the input data
+        // invalid and return an io::Error to the user.
+        if !self.stack.is_empty() || self.stack_str_size != 0 {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidData,
+                "Input data ends in the middle of a stack.",
+            ));
+        }
         Ok(())
     }
 
-    fn is_applicable(&mut self, input: &str) -> Option<bool> {
+    fn is_applicable_(&mut self, input: &str) -> Option<bool> {
         let mut found_empty_line = false;
         let mut found_stack_line = false;
         let mut input = input.as_bytes();
@@ -141,203 +159,93 @@ impl Collapse for Folder {
                 }
             }
         }
-
         None
     }
 
-    #[cfg(test)]
+    // This method should do the same thing as:
+    // ```
+    // fn is_end_of_stack(&self, line: &[u8]) -> bool {
+    //     match std::str::from_utf8(line) {
+    //         Ok(line) => {
+    //             let line = line.trim();
+    //             match line.parse::<usize>() {
+    //                 Ok(_) => true,
+    //                 Err(_) => false,
+    //             }
+    //         }
+    //         Err(_) => false,
+    //     }
+    // }
+    // ```
+    // But it is much faster since it works directly on bytes and because all we're interested in is
+    // whether the provided bytes **can** be parsed into a `usize`, not which `usize` they parse into.
+    // Also, we don't need to validate that the input is utf8.
+    //
+    // Benchmarking results for the two methods:
+    // * Using the method above: 281 MiB/s
+    // * Using the method below: 437 MiB/s
+    //
+    fn is_end_of_stack(&self, line: &[u8]) -> bool {
+        // In order to return `true`, as we iterate over the provided bytes, we need to progress
+        // through each of the follow states, in order; if we can't, immediately return `false`.
+        #[allow(clippy::enum_variant_names)]
+        enum State {
+            StartOfLine,  // Accept any number of whitespace characters
+            MiddleOfLine, // Accept any number of ascii digits
+            EndOfLine,    // Accept any number of whitespace characters
+        }
+        let mut state = State::StartOfLine;
+        for b in line {
+            let c = *b as char;
+            match state {
+                State::StartOfLine => {
+                    if c.is_whitespace() {
+                        continue;
+                    } else if c.is_ascii_digit() {
+                        state = State::MiddleOfLine;
+                    } else {
+                        return false;
+                    }
+                }
+                State::MiddleOfLine => {
+                    if c.is_ascii_digit() {
+                        continue;
+                    } else if c.is_whitespace() {
+                        state = State::EndOfLine;
+                    } else {
+                        return false;
+                    }
+                }
+                State::EndOfLine => {
+                    if c.is_whitespace() {
+                        continue;
+                    } else {
+                        return false;
+                    }
+                }
+            }
+        }
+        true
+    }
+
+    fn nstacks_per_job(&self) -> usize {
+        self.nstacks_per_job
+    }
+
     fn set_nstacks_per_job(&mut self, n: usize) {
         self.nstacks_per_job = n;
     }
 
-    #[doc(hidden)]
-    fn set_nthreads(&mut self, n: usize) {
+    fn nthreads(&self) -> usize {
+        self.opt.nthreads
+    }
+
+    fn set_nthreads_(&mut self, n: usize) {
         self.opt.nthreads = n;
-        self.occurrences = Occurrences::new(n);
     }
 }
 
 impl Folder {
-    fn collapse_single_threaded<R>(&mut self, mut reader: R) -> io::Result<()>
-    where
-        R: io::BufRead,
-    {
-        let mut line = String::new();
-        loop {
-            line.clear();
-
-            if reader.read_line(&mut line)? == 0 {
-                break;
-            }
-
-            let line = line.trim();
-
-            if line.is_empty() {
-                continue;
-            } else if let Ok(count) = line.parse::<usize>() {
-                self.on_stack_end(count);
-            } else {
-                self.on_stack_line(line);
-            }
-        }
-
-        if !self.stack.is_empty() || self.stack_str_size != 0 {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                "Input data ends in the middle of a stack.",
-            ));
-        }
-
-        Ok(())
-    }
-
-    fn collapse_multi_threaded<R>(&mut self, mut reader: R) -> io::Result<()>
-    where
-        R: io::BufRead,
-    {
-        assert_ne!(self.nstacks_per_job, 0);
-        assert!(self.occurrences.is_concurrent());
-        assert!(self.opt.nthreads > 1);
-
-        crossbeam::thread::scope(|scope| {
-            // Spin up the threadpool / worker threads.
-
-            // Channel for the main thread to send input data to the worker threads
-            let (tx_input, rx_input) = channel::bounded::<Option<Vec<u8>>>(2 * self.opt.nthreads);
-
-            // Channel for the worker threads to send an error, if any, to the main thread.
-            let (tx_output, rx_output) = channel::bounded::<io::Result<()>>(1);
-
-            // Channel for the worker threads to send a signal to the other worker threads that
-            // they have errored and so everyone should stop doing unnecessary work and return.
-            let (tx_stop, rx_stop) = channel::bounded::<()>(self.opt.nthreads - 1);
-
-            let mut handles = Vec::with_capacity(self.opt.nthreads);
-            for _ in 0..self.opt.nthreads {
-                let tx_output = tx_output.clone();
-                let rx_input = rx_input.clone();
-                let (tx_stop, rx_stop) = (tx_stop.clone(), rx_stop.clone());
-
-                let nstacks_per_job = self.nstacks_per_job;
-                let occurrences = self.occurrences.clone();
-                let opt = self.opt.clone();
-
-                let handle = scope.spawn(move |_| {
-                    let mut folder = Folder {
-                        cache_inlines: Vec::default(),
-                        nstacks_per_job,
-                        occurrences,
-                        stack: VecDeque::default(),
-                        stack_str_size: 0,
-                        opt,
-                    };
-                    // Loop until the main thread signals there is no more data
-                    // or another worker signals they have errored...
-                    loop {
-                        channel::select! {
-                            recv(rx_input) -> input => {
-                                let data = match input.unwrap() {
-                                    Some(data) => data,
-                                    None => return,
-                                };
-                                if let Err(e) = folder.collapse_single_threaded(&data[..]) {
-                                    // If we produce an error...
-
-                                    // Signal all the other threads; so they can stop doing needless work.
-                                    // (note: if this fails, it means another thread has already sent
-                                    // stop signals to all the other threads; so we don't have to.)
-                                    for _ in 0..(folder.opt.nthreads - 1) {
-                                        let _ = tx_stop.try_send(());
-                                    }
-
-                                    // Send the error back to the main thread for propagation.
-                                    // (note: if this fails it means another thread has already sent an error
-                                    // to the main thread; so we don't have to.)
-                                    let _ = tx_output.try_send(Err(e));
-
-                                    // Return.
-                                    return;
-                                }
-                            },
-                            recv(rx_stop) -> _ => return,
-                        }
-                    }
-                });
-                handles.push(handle);
-            }
-
-            drop(rx_input);
-
-            let buf_capacity =
-                usize::next_power_of_two(collapse::NBYTES_PER_STACK_GUESS * self.nstacks_per_job);
-            let mut buf = Vec::with_capacity(buf_capacity);
-            let (mut index, mut nstacks) = (0, 0);
-
-            // Loop through the input data in order to chunk it up and send it off to the worker threads...
-            loop {
-                // First, read some data into the buffer.
-                let n = reader.read_until(b'\n', &mut buf)?;
-
-                // If we're at the end of the data...
-                if n == 0 {
-                    // Send the last slice and exit the loop.
-                    let _ = tx_input.send(Some(buf));
-                    break;
-                }
-
-                let line = &buf[index..index + n];
-                index += n;
-
-                // If we're at the end of a stack...
-                if is_end_of_stack(line) {
-                    // Count it.
-                    nstacks += 1;
-                    // If we've seen enough stacks to make up a slice...
-                    if nstacks == self.nstacks_per_job {
-                        // Send it.
-                        let buf_capacity = buf.capacity();
-                        let chunk = mem::replace(&mut buf, Vec::with_capacity(buf_capacity));
-                        if tx_input.send(Some(chunk)).is_err() {
-                            // If this send returns an error, it means the children have shutdown
-                            // because one of them produced an error and has already sent
-                            // it to the main thread; so we should stop doing unnecessary work,
-                            // i.e. we should break.
-                            break;
-                        }
-                        // Reset the state; mark the beginning of the next slice.
-                        index = 0;
-                        nstacks = 0;
-                    }
-                }
-            }
-
-            // We've finished sending the input; so send the "no more input" signal,
-            // i.e. `None`, to the children.
-            for _ in &handles {
-                if tx_input.send(None).is_err() {
-                    // If this send returns an error, it means the children have shutdown
-                    // because one of them has already produced an error and sent it to the
-                    // main thread; so we should break here.
-                    break;
-                }
-            }
-
-            drop(tx_output);
-            let maybe_error = rx_output.iter().find(|result| result.is_err());
-
-            for handle in handles {
-                handle.join().unwrap();
-            }
-
-            if let Some(error) = maybe_error {
-                error?;
-            }
-
-            Ok(())
-        })
-        .unwrap()
-    }
-
     // This function approximates the Perl regex s/(::.*)[(<].*/$1/
     // from https://github.com/brendangregg/FlameGraph/blob/1b1c6deede9c33c5134c920bdb7a44cc5528e9a7/stackcollapse.pl#L88
     fn uncpp(probe: &str) -> &str {
@@ -411,7 +319,7 @@ impl Folder {
 
     // DTrace doesn't properly demangle Rust function names, so fix those.
     fn fix_rust_symbol<'a>(&self, frame: &'a str) -> Cow<'a, str> {
-        self.transform_function_name(frame, fix_partially_demangled_rust_symbol)
+        self.transform_function_name(frame, common::fix_partially_demangled_rust_symbol)
     }
 
     // we have a stack line that shows one stack entry from the preceeding event, like:
@@ -465,7 +373,7 @@ impl Folder {
         }
     }
 
-    fn on_stack_end(&mut self, count: usize) {
+    fn on_stack_end(&mut self, count: usize, occurrences: &mut Occurrences) {
         // allocate a string that is long enough to hold the entire stack string
         let mut stack_str = String::with_capacity(self.stack_str_size);
 
@@ -487,78 +395,12 @@ impl Folder {
         }
 
         // count it!
-        self.occurrences.add(stack_str, count);
+        occurrences.insert_or_add(stack_str, count);
 
         // reset for the next event
         self.stack_str_size = 0;
         self.stack.clear();
     }
-}
-
-// This function should do the same thing as:
-// ```
-// fn is_end_of_stack(line: &[u8]) -> bool {
-//     match std::str::from_utf8(line) {
-//         Ok(line) => {
-//             let line = line.trim();
-//             match line.parse::<usize>() {
-//                 Ok(_) => true,
-//                 Err(_) => false,
-//             }
-//         }
-//         Err(_) => false,
-//     }
-// }
-// ```
-// But it is much faster since it works directly on bytes and because all we're interested in is
-// whether the provided bytes **can** be parsed into a `usize`, not which `usize` they parse into.
-// Also, we don't need to validate that the input is utf8.
-//
-// Benchmarking results for the two methods:
-// * Using the method above: 281 MiB/s
-// * Using the method below: 437 MiB/s
-//
-fn is_end_of_stack(line: &[u8]) -> bool {
-    // In order to return `true`, as we iterate over the provided bytes, we need to progress
-    // through each of the follow states, in order; if we can't, immediately return `false`.
-    #[allow(clippy::enum_variant_names)]
-    enum State {
-        StartOfLine,  // Accept any number of whitespace characters
-        MiddleOfLine, // Accept any number of ascii digits
-        EndOfLine,    // Accept any number of whitespace characters
-    }
-    let mut state = State::StartOfLine;
-    for b in line {
-        let c = *b as char;
-        match state {
-            State::StartOfLine => {
-                if c.is_whitespace() {
-                    continue;
-                } else if c.is_ascii_digit() {
-                    state = State::MiddleOfLine;
-                } else {
-                    return false;
-                }
-            }
-            State::MiddleOfLine => {
-                if c.is_ascii_digit() {
-                    continue;
-                } else if c.is_whitespace() {
-                    state = State::EndOfLine;
-                } else {
-                    return false;
-                }
-            }
-            State::EndOfLine => {
-                if c.is_whitespace() {
-                    continue;
-                } else {
-                    return false;
-                }
-            }
-        }
-    }
-    true
 }
 
 #[cfg(test)]
@@ -571,7 +413,8 @@ mod tests {
     use rand::{Rng, SeedableRng};
 
     use super::*;
-    use crate::collapse::tests_common;
+    use crate::collapse::common;
+    use crate::collapse::Collapse;
 
     lazy_static! {
         static ref INPUT: Vec<PathBuf> = {
@@ -611,7 +454,7 @@ mod tests {
     #[test]
     fn test_collapse_multi_dtrace() -> io::Result<()> {
         let mut folder = Folder::default();
-        tests_common::test_collapse_multi(&mut folder, &INPUT)
+        common::testing::test_collapse_multi(&mut folder, &INPUT)
     }
 
     #[test]
@@ -666,7 +509,7 @@ mod tests {
     #[ignore]
     fn bench_nstacks_dtrace() -> io::Result<()> {
         let mut folder = Folder::default();
-        tests_common::bench_nstacks(&mut folder, &INPUT)
+        common::testing::bench_nstacks(&mut folder, &INPUT)
     }
 
     #[test]
@@ -683,7 +526,7 @@ mod tests {
         let mut buf_expected = Vec::new();
         let mut count = 0;
 
-        let inputs = tests_common::read_inputs(&INPUT)?;
+        let inputs = common::testing::read_inputs(&INPUT)?;
 
         loop {
             let nstacks_per_job = rng.gen_range(1, 500 + 1);
