+++ src/collapse/guess.rs
@@ -27,6 +27,7 @@ impl Default for Options {
 ///
 /// If no applicable collapser is found, an error will be logged and
 /// nothing will be written.
+#[derive(Clone)]
 pub struct Folder {
     opt: Options,
 }
@@ -59,7 +60,11 @@ impl Collapse for Folder {
             options.nthreads = self.opt.nthreads;
             perf::Folder::from(options)
         };
-        let mut sample = sample::Folder::from(sample::Options::default());
+        let mut sample = {
+            let mut options = sample::Options::default();
+            options.nthreads = self.opt.nthreads;
+            sample::Folder::from(options)
+        };
 
         // Each Collapse impl gets its own flag in this array.
         // It gets set to true when the impl has been ruled out.
@@ -111,64 +116,7 @@ impl Collapse for Folder {
         unreachable!()
     }
 
-    #[cfg(test)]
-    fn set_nstacks_per_job(&mut self, _: usize) {}
-
-    #[doc(hidden)]
     fn set_nthreads(&mut self, n: usize) {
         self.opt.nthreads = n;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use lazy_static::lazy_static;
-
-    use super::*;
-    use crate::collapse::tests_common;
-
-    lazy_static! {
-        static ref INPUT: Vec<PathBuf> = {
-            [
-                // perf
-                "./flamegraph/example-perf-stacks.txt.gz",
-                "./flamegraph/test/perf-cycles-instructions-01.txt",
-                "./flamegraph/test/perf-dd-stacks-01.txt",
-                "./flamegraph/test/perf-funcab-cmd-01.txt",
-                "./flamegraph/test/perf-funcab-pid-01.txt",
-                "./flamegraph/test/perf-iperf-stacks-pidtid-01.txt",
-                "./flamegraph/test/perf-java-faults-01.txt",
-                "./flamegraph/test/perf-java-stacks-01.txt",
-                "./flamegraph/test/perf-java-stacks-02.txt",
-                "./flamegraph/test/perf-js-stacks-01.txt",
-                "./flamegraph/test/perf-mirageos-stacks-01.txt",
-                "./flamegraph/test/perf-numa-stacks-01.txt",
-                "./flamegraph/test/perf-rust-Yamakaky-dcpu.txt",
-                "./flamegraph/test/perf-vertx-stacks-01.txt",
-                "./tests/data/collapse-perf/empty-line.txt",
-                "./tests/data/collapse-perf/go-stacks.txt",
-                "./tests/data/collapse-perf/java-inline.txt",
-                "./tests/data/collapse-perf/weird-stack-line.txt",
-                // dtrace
-                "./flamegraph/example-dtrace-stacks.txt",
-                "./tests/data/collapse-dtrace/flamegraph-bug.txt",
-                "./tests/data/collapse-dtrace/hex-addresses.txt",
-                "./tests/data/collapse-dtrace/java.txt",
-                "./tests/data/collapse-dtrace/only-header-lines.txt",
-                "./tests/data/collapse-dtrace/scope_with_no_argument_list.txt",
-
-            ]
-            .into_iter()
-            .map(PathBuf::from)
-            .collect::<Vec<_>>()
-        };
-    }
-
-    #[test]
-    fn test_collapse_multi_guess() -> io::Result<()> {
-        let mut folder = Folder::default();
-        tests_common::test_collapse_multi(&mut folder, &INPUT)
     }
 }
